Rename files, don't use the `community` prefix
