+++ apps/researcher/src/app/[locale]/(objects)/objects/[id]/object-lists-actions.ts
@@ -1,6 +1,6 @@
 'use server';
 
-import {getCommunityBySlug} from '@/lib/community/community-actions';
+import {getCommunityBySlug} from '@/lib/community/actions';
 import {objectList} from '@colonial-collections/database';
 import {ObjectItemBeingCreated} from '@colonial-collections/database';
 import {revalidatePath} from 'next/cache';
