+++ apps/researcher/src/app/[locale]/communities/layout.tsx
@@ -1,5 +1,5 @@
 import {ListProvider} from '@colonial-collections/list-store';
-import {defaultSortBy} from '@/lib/community/community-actions';
+import {defaultSortBy} from '@/lib/community/actions';
 import {ReactNode} from 'react';
 
 export default function ObjectLayout({children}: {children: ReactNode}) {
